ECC-1467: Simplification of templates
