+++ src/grib_accessor_class_data_raw_packing.cc
@@ -11,6 +11,7 @@
  *  Enrico Fucile
  ****************************/
 
+#include "grib_api_internal.h"
 #include "grib_api_internal_cpp.h"
 
 #define PRE_PROCESSING_NONE 0
