+++ src/grib_api_internal_cpp.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "grib_accessor.h"
 #include "grib_value.h"
 #include "grib_bits_any_endian_simple.h"
 #include "grib_ieeefloat.h"
